+++ src/main/resources/templates/admin/index.html
@@ -27,13 +27,13 @@
         </p>
         <ul class="text-blue-700">
           <li class="inline-block p-1">
-            <a href="/stations" class="underline">역 관리</a>
+            <a href="/admin/stations" class="underline">역 관리</a>
           </li>
           <li class="inline-block p-1">
-            <a href="/lines" class="underline">노선 관리</a>
+            <a href="/admin/lines" class="underline">노선 관리</a>
           </li>
           <li class="inline-block p-1">
-            <a href="/edges" class="underline">구간 관리</a>
+            <a href="/admin/edges" class="underline">구간 관리</a>
           </li>
         </ul>
       </div>
